LayoutTests/fast/regex/slow.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
</head>
<body>
<script>
description(
'Test for expressions that would hang when evaluated due to exponential matching behavior. If the test does not hang it is a success.'
);

shouldBe('/(?:[^(?!)]||){23}z/.test("/(?:[^(?!)]||){23}z/")', 'false');
</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
